Refactor: Asset 압축
