+++ src/components/LoginHeader.tsx
@@ -1,4 +1,4 @@
-import icon from "../assets/domisaHeartIcon.svg";
+import icon from "../assets/domisaHeartIcon.png";
 import { useNavigate } from "react-router-dom";
 
 function LoginHeader() {
