+++ src/components/homePageHeader.tsx
@@ -1,4 +1,4 @@
-import icon from "../assets/domisaHeartIcon.svg"
+import icon from "../assets/domisaHeartIcon.png"
 import { useNavigate } from "react-router-dom";
 
 const festivalStart = { year: 2026, month: 5, day: 13 };
