+++ src/pages/HomePage.tsx
@@ -3,9 +3,9 @@ import nightBgImg from "../assets/homePageNightBackGround.jpg";
 import Header from "../components/homePageHeader";
 import MessageSlider from "../components/MessageSlider";
 import logo from "../assets/domisaLogo.svg";
-import dogImg from "../assets/dogIcon.svg";
-import heartImg from "../assets/domisaHeartIcon.svg";
-import catImg from "../assets/catIcon.svg";
+import dogImg from "../assets/dogIcon.png";
+import heartImg from "../assets/domisaHeartIcon.png";
+import catImg from "../assets/catIcon.png";
 import dayMapImg from "../assets/dayMapIcon.png";
 import nightMapImg from "../assets/nightMapIcon.png";
 import arrowImg from "../assets/arrowIcon.svg";
