+++ src/pages/MyPage/MyPage.tsx
@@ -10,9 +10,9 @@ import RightArrow from '../../assets/right_arrow.svg?react';
 import WithdrawConfirmModal from './WithdrawConfirmModal';
 import editPencilImg from '../../assets/edit_pencil.svg';
 import cookieImg from '../../assets/cookie.svg';
-import dogImg from '../../assets/dogIcon.svg';
-import domisaHeartImg from '../../assets/domisaHeartIcon.svg';
-import catImg from '../../assets/catIcon.svg';
+import dogImg from '../../assets/dogIcon.png';
+import domisaHeartImg from '../../assets/domisaHeartIcon.png';
+import catImg from '../../assets/catIcon.png';
 import flowerImg from '../../assets/flowerIcon.svg';
 import arrowIcon from '../../assets/arrowIcon.svg';
 import heartIconOrange from '../../assets/heartIconOrange.svg';
